Utils - Merging table and form files
Since we have 3 main sections (server, form and modals), we can insert table functions into the form file.
