+++ utils/form.py
@@ -2,6 +2,9 @@
 Utils functions to retrieve data from QGIS attributes from
 """
 
+from qgis.PyQt.QtWidgets import QTableWidgetItem
+from qgis.PyQt.QtCore import Qt
+
 def format_map(widget_map):
     """
     Returns a JSON-like version of a ValueMap widget's map option
@@ -71,3 +74,42 @@ def layers_configs(layers):
         {config["name"]: config for config in flat_configs}.values()
     )
 
+def create_label_item(title):
+    item = QTableWidgetItem(title)
+    item.setTextAlignment(Qt.AlignCenter)
+    item.setFlags(Qt.NoItemFlags)
+    return item
+
+def create_checkbox_item(attribute):
+    item = QTableWidgetItem()
+    item.setFlags(Qt.ItemIsUserCheckable | Qt.ItemIsEnabled)
+    item.setCheckState(Qt.Unchecked)
+    item.setData(1, attribute)
+    return item
+
+def create_row(field):
+    return dict(
+        label=create_label_item(field["name"]),
+        checkbox_hidden=create_checkbox_item('hidden'),
+        checkbox_disabled=create_checkbox_item('disabled')
+    )
+
+
+def create_rows(fields):
+    return list(map(lambda f: create_row(f), fields))
+
+
+def fill_table(table, fields):
+    rows = create_rows(fields)
+    for index, row in enumerate(rows):
+        table.setItem(index, 0, row["label"])
+        table.setItem(index, 1, row["checkbox_disabled"])
+        table.setItem(index, 2, row["checkbox_hidden"])
+
+def fields_display(table):
+    res = []
+    for row in range(table.rowCount()):
+        field_name = table.item(row, 0).text()
+        disabled, hidden = table.item(row, 1).checkState() == Qt.Checked, table.item(row, 2).checkState() == Qt.Checked
+        res.append(dict(field_name=field_name, disabled=disabled, hidden=hidden))
+    return res
